+++ src/Resources/Policy.Autorest/test/PolicyDefinitionCRUD.Tests.ps1
@@ -16,6 +16,10 @@ Describe 'PolicyDefinitionCRUD' -Tag 'LiveOnly' {
         $actual = Get-AzPolicyDefinition -Name $policyName
         $actual | Should -Not -BeNullOrEmpty
         $expected.Name | Should -Be $actual.Name
+        $actual.Version | Should -Be $defaultVersion
+        $expected.Version | Should -Be $actual.Version
+        $expected.Versions | Should -HaveCount 1
+        $expected.Versions | Should -Be $actual.Versions
         $expected.PolicyDefinitionId | Should -Be $actual.PolicyDefinitionId
         $actual.PolicyRule | Should -Not -BeNullOrEmpty
         $expected.Mode | Should -Be $actual.Mode
@@ -27,6 +31,10 @@ Describe 'PolicyDefinitionCRUD' -Tag 'LiveOnly' {
         $expected = Get-AzPolicyDefinition -Name $policyName
         $expected.DisplayName | Should -Be $actual.DisplayName
         $expected.Description | Should -Be $actual.Description
+        $actual.Version | Should -Be $defaultVersion
+        $expected.Version | Should -Be $actual.Version
+        $expected.Versions | Should -Be $actual.Versions
+        $actual.Versions | Should -Not -BeNullOrEmpty
         $actual.Metadata | Should -Not -BeNullOrEmpty
         $actual.Metadata.$metadataName | Should -Be $metadataValue
     }
@@ -44,10 +52,27 @@ Describe 'PolicyDefinitionCRUD' -Tag 'LiveOnly' {
         $actual = Get-AzPolicyDefinition -Name $policyName
         $actual.DisplayName | Should -Be $expected.DisplayName
         $actual.Description | Should -Be $expected.Description
+        $expected.Version | Should -Be $defaultVersion
+        $expected.Version | Should -Be $actual.Version
+        $expected.Versions | Should -Be $actual.Versions
         $actual.Metadata.$metadataName | Should -Be $expected.Metadata.$metadataName
         $actual.PolicyDefinition | Should -BeLike $expected.PolicyDefinition
         $actual.Parameter | Should -BeLike $expected.Parameter
         $actual.Mode | Should -BeLike $expected.Mode
+    }
+
+    It 'Make a policy definition with version from rule file' {
+        # make a policy definition with version, get it back and validate
+        $expected = New-AzPolicyDefinition -Name $policyName -Policy "$testFilesFolder\SamplePolicyDefinition.json" -Mode Indexed -Description $description -Version $someNewVersion
+        $actual = Get-AzPolicyDefinition -Name $policyName
+        $actual | Should -Not -BeNullOrEmpty
+        $actual.Version | Should -Be $someNewVersion
+        $actual.Versions | Should -HaveCount 2
+        $expected.Name | Should -Be $actual.Name
+        $expected.Version | Should -Be $actual.Version
+        $expected.PolicyDefinitionId | Should -Be $actual.PolicyDefinitionId
+        $actual.PolicyRule | Should -Not -BeNullOrEmpty
+        $expected.Mode | Should -Be $actual.Mode
     }
 
     It 'Make policy definition from command line rule' {